dashboard/dashboard-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.129.0"}
psycopg = {extras = ["binary"], version = "^3.3.2"}
pyjwt = "^2.11.0"
passlib = "^1.7.4"
requests = "^2.32.5"
cryptography = "^46.0.5"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
